pyproject.toml
[tool.poetry]
name = "housing-market-dashboard"
version = "0.1.0"
description = ""
authors = ["tcstrength <thai.cuong1404@gmail.com>"]
readme = "README.md"
packages = [{include="hmd", from="src"}]

[tool.poetry.dependencies]
python = "^3.12"
docker-py = "^1.10.6"
selenium = "^4.25.0"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
matplotlib = "^3.9.2"
lxml = "^5.3.0"
pydantic = "^2.9.2"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
pandas = "^2.2.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
